requirements.txt
Authlib==1.6.5
deepface==0.0.79
eventlet==0.33.3
fastapi==0.104.1
fer==22.5.1
Flask==3.1.2
flask-cors==4.0.0
flask-socketio==5.3.5
gTTS==2.4.0
imageio==2.33.0
imageio-ffmpeg==0.4.9
keras==2.15.0
marshmallow==3.20.1
nltk==3.9.2
numpy==1.26.4
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
openai==1.3.0
pandas
pdfplumber==0.11.7
Pillow==12.0.0
pyaudio==0.2.14
pydub==0.25.1
pypdf
python-docx
python-dotenv==1.1.1
python-socketio==5.10.0
pyttsx3==2.90
reportlab==4.4.4
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers
SpeechRecognition==3.10.0
sqlite-utils==3.36
tensorflow==2.15.0
textblob==0.19.0
tqdm==4.67.1
uvicorn==0.24.0